frontend plus szukajka

+++ src/main/java/pl/sda/grocefy/product/dto/ItemDTO.java
@@ -15,6 +15,6 @@ public class ItemDTO {
     private Long id;
     private String product;
     private String hash;
-    private int count;
+    private Double count;
     private Unit unit;
 }
